# service/bot/openapi/bot.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: bot
  description: handles webhook messages and executes plugins based on message content
servers:
  - url: https://{hostname}
    variables:
      hostname:
        default: localhost
paths:          
  /pubsub/subscriptions/video-saved:
    post:
      operationId: handleVideoSavedMessage
      requestBody:
        description: handles messages from video saved topic
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PubSubMessage"
      responses:
        "200":
          description: video was sucessfully published to telegram
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /pubsub/subscriptions/video-save-failed:
    post:
      operationId: handleVideoSaveFailedMessage
      requestBody:
        description: handles messages from dead letter topic
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PubSubMessage"
      responses:
        "200":
          description: failure message was sucessfully published to telegram
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /telegram/callback:
    post:
      operationId: handleTelegramMessage
      requestBody:
        description: handles messages from telegram webhook
        required: true
        content:
          application/json:
           schema:
            type: object
            nullable: true
      responses:
        "200":
          description: message was handled
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    PubSubMessage:
      type: object
      required: [message]
      properties:
        message:
          type: object
          required: [message_id, data, publish_time]
          properties:
            message_id:
              type: string
            data:
              type: string
            publish_time:
              type: string
              format: date-time

    Error:
      type: object
      required:
        - slug
        - message
      properties:
        slug:
          type: string
        message:
          type: string